requirements.txt
Flask~=3.1.0
requests~=2.31.0
openai~=0.27.0
pymongo~=4.12.1
sshtunnel~=0.4.0
datetime~=5.5
apscheduler~=3.10.1
dataclasses~=0.6
gspread~=6.2.1
google-auth~=2.40.3